jarvis-whatsapp-assistant/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1
aiohttp==3.9.1
openai==1.3.7
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
httpx==0.25.2
python-dotenv==1.0.0
